# openapi/swagger/swagger.yaml
basePath: /
definitions:
...
  main.SwaggerEnvelope:
    properties:
      data: {}
      error:
        type: string
      pagination:
        $ref: '#/definitions/main.SwaggerPagination'
      success:
        type: boolean
    type: object
  main.SwaggerPagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  model.AdvertisementCreateReq:
    properties:
      image:
        type: string
      imageHeight:
        type: integer
      imageWidth:
        type: integer
      isActive:
        type: boolean
      link:
        type: string
      position:
        type: string
      title:
        type: string
    type: object
  model.AdvertisementPositionDTO:
    properties:
      _count:
        additionalProperties:
          type: integer
        type: object
      position:
        type: string
    type: object
  model.AdvertisementResp:
    properties:
      createdAt:
        type: string
      id:
        type: string
      image:
        type: string
      imageHeight:
        type: integer
      imageWidth:
        type: integer
      isActive:
        type: boolean
      link:
        type: string
      position:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  model.AdvertisementStatsResp:
    properties:
      activeAds:
        type: integer
      inactiveAds:
        type: integer
      positionStats:
        items:
          $ref: '#/definitions/model.AdvertisementPositionDTO'
        type: array
      totalAds:
        type: integer
    type: object
  model.AdvertisementUpdateReq:
    properties:
      image:
        type: string
      imageHeight:
        type: integer
      imageWidth:
        type: integer
      isActive:
        type: boolean
      link:
        type: string
      position:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: Contract for the Nex HTTP API.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  title: Nex API
  version: 1.0.0
paths:
  /api/advertisements:
    get:
      operationId: advertisements_route_get
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - description: position
        in: query
        name: position
        type: string
      - description: isActive
        in: query
        name: isActive
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.AdvertisementResp'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: GET /api/advertisements
      tags:
      - advertisements
    post:
      consumes:
      - application/json
      operationId: advertisements_route_post
      parameters:
      - description: request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AdvertisementCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.AdvertisementResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: POST /api/advertisements
      tags:
      - advertisements
  /api/advertisements/{id}:
    delete:
      operationId: advertisements_id_route_delete
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: DELETE /api/advertisements/{id}
      tags:
      - advertisements
    get:
      operationId: advertisements_id_route_get
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.AdvertisementResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: GET /api/advertisements/{id}
      tags:
      - advertisements
    put:
      consumes:
      - application/json
      operationId: advertisements_id_route_put
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.AdvertisementUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.AdvertisementResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: PUT /api/advertisements/{id}
      tags:
      - advertisements
  /api/advertisements/{id}/toggle:
    put:
      operationId: advertisements_id_toggle_route_put
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.AdvertisementResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: PUT /api/advertisements/{id}/toggle
      tags:
      - advertisements
  /api/advertisements/by-position/{position}:
    get:
      operationId: advertisements_by_position_position_route_get
      parameters:
      - description: position
        in: path
        name: position
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.AdvertisementResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: GET /api/advertisements/by-position/{position}
      tags:
      - advertisements
  /api/advertisements/stats:
    get:
      operationId: advertisements_stats_route_get
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.AdvertisementStatsResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: GET /api/advertisements/stats
      tags:
      - advertisements
...
securityDefinitions:
  ApiTokenAuth:
    in: header
    name: Authorization
    type: apiKey
  CronSecretAuth:
    in: header
    name: Authorization
    type: apiKey
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"